requirements.txt
streamlit
openai
sqlalchemy
pymysql
python-dotenv
psycopg2-binary
pandas
langchain
tqdm
cryptography

